+++ src/test/php/webservices/rest/unittest/IssuesWithField.class.php
@@ -40,6 +40,6 @@ class IssuesWithField extends \lang\Object {
    * @return  string
    */
   public function toString() {
-    return $this->getClassName().'@'.\xp::stringOf($this->issues);
+    return nameof($this).'@'.\xp::stringOf($this->issues);
   }
 }